refactor: move Playwright specs into their example dirs (rf2-kdvh)
examples/playwright/ was misnamed: it housed Playwright spec files for
the example apps but read like a peer example called 'playwright'.

Each *.spec.cjs is now relocated to live alongside the example it
exercises:

  examples/counter/counter.spec.cjs
  examples/todomvc/todomvc.spec.cjs
  examples/login/login.spec.cjs
  examples/routing/routing.spec.cjs
  examples/realworld/realworld.spec.cjs
  examples/nine_states/nine_states.spec.cjs
  examples/managed_http_counter/managed_http_counter.spec.cjs
  examples/seven_guis/{cells,circle_drawer,crud,flight_booker,temperature,timer}.spec.cjs

The shared assertion helpers (formerly _helpers.cjs) move to
implementation/scripts/spec-helpers.cjs alongside the runner that
uses them; each spec's require path is updated accordingly.

Pure relocation: no semantic changes to spec bodies. Orchestrator
spec discovery is updated separately.
