+++ examples/login/login.spec.cjs
@@ -15,7 +15,7 @@
  *   - On success the banner switches to "Welcome!".
  */
 
-const { expectVisible, expectTextContains } = require('./_helpers.cjs');
+const { expectVisible, expectTextContains } = require('../../implementation/scripts/spec-helpers.cjs');
 
 module.exports = {
   name: 'login (state-machine demo)',
